requirements.txt
Flask-API.yandex
PyJWT
sqlalchemy
flask
